third_party/blink/web_tests/platform/generic/accessibility/presentational-elements-with-focus-expected.txt











This tests that the ARIA presentation role is ignored when an element can set focus.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS accessibilityController.focusedElement.childrenCount is 12
PASS successfullyParsed is true

TEST COMPLETE
